+++ app/build.gradle.kts
@@ -55,7 +55,8 @@ android {
 dependencies {
 
     // modules
-    implementation(projects.commonUi)
+    implementation(projects.coreDesignsystem)
+    implementation(projects.coreUi)
     implementation(projects.coreModel)
     implementation(projects.coreData)
     implementation(projects.featurePopular)
