--- common-ui/build.gradle.kts
@@ -1,69 +1,0 @@
-@Suppress("DSL_SCOPE_VIOLATION")
-plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.kotlin.android)
-}
-
-android {
-    namespace = "com.jik.common.ui"
-    compileSdk = 33
-
-    defaultConfig {
-        minSdk = 24
-        targetSdk = 33
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-    }
-    buildFeatures {
-        compose = true
-    }
-    composeOptions {
-        kotlinCompilerExtensionVersion = "1.4.3"
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_1_8
-        targetCompatibility = JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
-
-dependencies {
-
-    // modules
-    implementation(projects.coreModel)
-
-    implementation(libs.androidx.ktx)
-
-    // compose
-    implementation(libs.compose.ui)
-    implementation(libs.compose.material3)
-    implementation(libs.compose.tooling)
-    implementation(libs.compose.preview)
-
-    // coroutines
-    implementation(libs.coroutines)
-
-    // coil
-    implementation(libs.compose.coil)
-
-    // test
-    testImplementation(libs.junit)
-    testImplementation(libs.coroutines.test)
-
-    // android test
-    androidTestImplementation(libs.ext.junit)
-    androidTestImplementation(libs.espresso)
-}
