--- common-ui/src/main/java/com/jik/common/ui/preview/DevicePreviews.kt
@@ -1,9 +1,0 @@
-package com.jik.common.ui.preview
-
-import androidx.compose.ui.tooling.preview.Preview
-
-@Preview(name = "phone", device = "spec:shape=Normal,width=360,height=640,unit=dp,dpi=480")
-@Preview(name = "landscape", device = "spec:shape=Normal,width=640,height=360,unit=dp,dpi=480")
-@Preview(name = "foldable", device = "spec:shape=Normal,width=673,height=841,unit=dp,dpi=480")
-@Preview(name = "tablet", device = "spec:shape=Normal,width=1280,height=800,unit=dp,dpi=480")
-annotation class DevicePreviews
