--- common-ui/src/main/java/com/jik/common/ui/util/Toast.kt
@@ -1,8 +1,0 @@
-package com.jik.common.ui.util
-
-import android.content.Context
-import android.widget.Toast
-
-fun toast(context: Context, message: String) {
-    Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
-}
