+++ core-designsystem/build.gradle.kts
@@ -2,12 +2,10 @@
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.kotlin.kapt)
-    id("dagger.hilt.android.plugin")
 }
 
 android {
-    namespace = "com.jik.feature.popular"
+    namespace = "com.jik.core.designsystem"
     compileSdk = 33
 
     defaultConfig {
@@ -15,6 +13,7 @@ android {
         targetSdk = 33
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles("consumer-rules.pro")
     }
 
     buildTypes {
@@ -26,6 +25,12 @@ android {
             )
         }
     }
+    buildFeatures {
+        compose = true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion = "1.4.3"
+    }
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_1_8
         targetCompatibility = JavaVersion.VERSION_1_8
@@ -33,48 +38,32 @@ android {
     kotlinOptions {
         jvmTarget = "1.8"
     }
-    buildFeatures {
-        compose = true
-    }
-    composeOptions {
-        kotlinCompilerExtensionVersion = "1.4.3"
-    }
 }
 
 dependencies {
 
     // modules
-    implementation(projects.coreDesignsystem)
     implementation(projects.coreUi)
-    implementation(projects.coreModel)
-    implementation(projects.coreData)
 
     implementation(libs.androidx.ktx)
-    implementation(libs.androidx.runtime.ktx)
-    implementation(libs.androidx.activity)
 
     // compose
     implementation(libs.compose.ui)
-    implementation(libs.compose.preview)
     implementation(libs.compose.material3)
+    implementation(libs.compose.tooling)
+    implementation(libs.compose.preview)
 
-    // lifecycle
-    implementation(libs.lifecycle.viewmodel)
-    implementation(libs.lifecycle.runtime)
+    // coroutines
+    implementation(libs.coroutines)
 
-    // hilt
-    implementation(libs.hilt.android)
-    kapt(libs.hilt.compiler)
+    // coil
+    implementation(libs.compose.coil)
 
     // test
     testImplementation(libs.junit)
+    testImplementation(libs.coroutines.test)
 
     // android test
     androidTestImplementation(libs.ext.junit)
     androidTestImplementation(libs.espresso)
-    androidTestImplementation(libs.compose.test.junit)
-
-    // debug
-    debugImplementation(libs.compose.tooling)
-    debugImplementation(libs.compose.test.manifest)
 }
