+++ core-designsystem/src/main/AndroidManifest.xml
