+++ core-ui/src/main/java/com/jik/core/ui/util/Toast.kt
@@ -1,0 +1,8 @@
+package com.jik.core.ui.util
+
+import android.content.Context
+import android.widget.Toast
+
+fun toast(context: Context, message: String) {
+    Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
+}
