+++ feature-popular/src/main/java/com/jik/feature/popular/PopularViewModel.kt
@@ -2,10 +2,10 @@ package com.jik.feature.popular
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.jik.common.ui.UiState
-import com.jik.common.ui.getUiStateFlow
 import com.jik.core.data.repository.MovieRepository
 import com.jik.core.model.Movie
+import com.jik.core.ui.state.UiState
+import com.jik.core.ui.state.getUiStateFlow
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.StateFlow
 import javax.inject.Inject
